+++ src/Core/Util/LocaleCodeProvider.php
@@ -6,7 +6,8 @@ use Psr\Log\LoggerInterface;
 use Shopware\Core\{
 	Framework\Context,
 	Framework\DataAbstractionLayer\EntityRepositoryInterface,
-	Framework\DataAbstractionLayer\Search\Criteria};
+	Framework\DataAbstractionLayer\Search\Criteria,
+	System\Language\LanguageCollection};
 
 /**
  * Class LocaleCodeProvider
@@ -64,5 +65,50 @@ class LocaleCodeProvider {
 		}
 
 		return $language->getLocale() ? $language->getLocale()->getCode() : $defaultLocale;
+	}
+
+	/**
+	 * Get available languages
+	 *
+	 * @param \Shopware\Core\Framework\Context $context
+	 * @return \Shopware\Core\System\Language\LanguageCollection
+	 */
+	public function getAvailableLanguages(Context $context): LanguageCollection
+	{
+		return $this->languageRepository->search((new Criteria())->addAssociations([
+			'locale',
+		]), $context)->getEntities();
+	}
+
+	/**
+	 * Get available translations
+	 *
+	 * @param array                            $translation
+	 * @param \Shopware\Core\Framework\Context $context
+	 *
+	 * @return array
+	 */
+	public function getAvailableTranslations(array $translation, Context $context): array
+	{
+		$availableLanguages = $this->getAvailableLanguages($context);
+
+		$locales = $locales = array_map(
+			function ($language) {
+				/**
+				 * @var \Shopware\Core\System\Language\LanguageEntity $language
+				 */
+				return $language->getLocale()->getCode();
+			},
+			$availableLanguages->jsonSerialize()
+		);
+
+		foreach ($translation as $key => $value) {
+			if (!in_array($key, $locales)) {
+				$translation[$key] = null;
+			}
+		}
+
+		$translation = array_filter($translation);
+		return $translation;
 	}
 }